+++ libmtx/mtxfile/comments.h
@@ -16,7 +16,7 @@
  * along with Libmtx.  If not, see <https://www.gnu.org/licenses/>.
  *
  * Authors: James D. Trotter <james@simula.no>
- * Last modified: 2022-01-04
+ * Last modified: 2022-04-14
  *
  * Matrix Market comment lines.
  */
@@ -142,7 +142,7 @@ int mtxfilecomments_printf(
 int mtxfile_fread_comments(
     struct mtxfilecomments * comments,
     FILE * f,
-    int * lines_read,
+    int64_t * lines_read,
     int64_t * bytes_read,
     size_t line_max,
     char * linebuf);
@@ -159,7 +159,7 @@ int mtxfile_fread_comments(
 int mtxfile_gzread_comments(
     struct mtxfilecomments * comments,
     gzFile f,
-    int * lines_read,
+    int64_t * lines_read,
     int64_t * bytes_read,
     size_t line_max,
     char * linebuf);
